Port Debian and Ubuntu to DistributionInstaller
